+++ internal/provider/data_source_sdwan_traffic_data_policy_definition_test.go
@@ -40,6 +40,7 @@ func TestAccDataSourceSdwanTrafficDataPolicyDefinition(t *testing.T) {
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.name", "Seq1"),
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.type", "applicationFirewall"),
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.ip_type", "ipv4"),
+					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.base_action", "accept"),
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.match_entries.0.type", "appList"),
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.match_entries.0.application_list_id", "e3aad846-abb9-425f-aaa8-9ed17b9c8d7c"),
 					resource.TestCheckResourceAttr("data.sdwan_traffic_data_policy_definition.test", "sequences.0.action_entries.0.type", "log"),
@@ -61,6 +62,7 @@ resource "sdwan_traffic_data_policy_definition" "test" {
     name = "Seq1"
     type = "applicationFirewall"
     ip_type = "ipv4"
+    base_action = "accept"
 	match_entries = [{
 		type = "appList"
 		application_list_id = "e3aad846-abb9-425f-aaa8-9ed17b9c8d7c"